+++ src/compojure/api/meta.clj
@@ -1,16 +1,13 @@
 (ns compojure.api.meta
-  (:require [compojure.api.common :refer :all]
+  (:require [clojure.walk :refer [keywordize-keys]]
+            [compojure.api.common :refer :all]
             [compojure.core :refer [routes]]
-            [ring.util.response :as response]
-            [ring.util.http-response :refer [internal-server-error]]
-            [plumbing.core :refer :all]
+            [plumbing.core :refer :all :exclude [update]]
             [plumbing.fnk.impl :as fnk-impl]
-            [ring.swagger.core :as swagger]
-            [ring.swagger.schema :as schema]
             [ring.swagger.common :refer :all]
-            [schema.core :as s]
-            [clojure.walk16 :refer [keywordize-keys]]
-            [clojure.tools.macro :refer [name-with-attributes]]))
+            [ring.swagger.schema :as schema]
+            [ring.util.http-response :refer [internal-server-error]]
+            [schema.core :as s]))
 
 (def +compojure-api-request+
   "lexically bound ring-request for handlers."
@@ -82,89 +79,90 @@
 ;;
 ;; Smart restructurings
 ;;
+
+; Defines a return type and coerced the return value of a body against it.
+; Examples:
+; :return MyModel
+; :return {:value String}
+; :return #{{:key (s/maybe Long)}}
 (defmethod restructure-param :return [k model acc]
-  "Defines a return type and coerced the return value of a body against it.
-   Examples:
-   :return MyModel
-   :return {:value String}
-   :return #{{:key (s/maybe Long)}}"
   (-> acc
       (update-in [:parameters] assoc k model)
       (update-in [:responses] assoc 200 model)))
 
+; value is a map of http-response-code -> Model. Translates to both swagger
+; parameters and return model coercion. Models can be decorated with meta-data.
+; Examples:
+; :responses {403 ErrorEnvelope}
+; :responses {403 ^{:message \"Underflow\"} ErrorEnvelope}
 (defmethod restructure-param :responses [_ responses acc]
-  "value is a map of http-response-code -> Model. Translates to both swagger
-   parameters and return model coercion. Models can be decorated with meta-data.
-   Examples:
-   :responses {403 ErrorEnvelope}
-   :responses {403 ^{:message \"Underflow\"} ErrorEnvelope}"
   (let [messages (responses->messages responses)]
     (-> acc
         (update-in [:parameters :responseMessages] (comp distinct concat) messages)
         (update-in [:responses] merge responses))))
 
+; reads body-params into a enchanced let. First parameter is the let symbol,
+; second is the Model to coerced! against.
+; Examples:
+; :body [user User]
 (defmethod restructure-param :body [_ [value model] acc]
-  "reads body-params into a enchanced let. First parameter is the let symbol,
-   second is the Model to coerced! against.
-   Examples:
-   :body [user User]"
   (-> acc
       (update-in [:lets] into [value (src-coerce! model :body-params :json)])
       (update-in [:parameters :parameters] conj {:type :body
                                                  :model model})))
 
+; reads query-params into a enchanced let. First parameter is the let symbol,
+; second is the Model to coerced! against.
+; Examples:
+; :query [user User]
 (defmethod restructure-param :query [_ [value model] acc]
-  "reads query-params into a enchanced let. First parameter is the let symbol,
-   second is the Model to coerced! against.
-   Examples:
-   :query [user User]"
   (-> acc
       (update-in [:lets] into [value (src-coerce! model :query-params :query)])
       (update-in [:parameters :parameters] conj {:type :query
                                                  :model model})))
 
+; restructures body-params with plumbing letk notation. Example:
+; :body-params [id :- Long name :- String]
 (defmethod restructure-param :body-params [_ body-params acc]
-  "restructures body-params with plumbing letk notation. Example:
-   :body-params [id :- Long name :- String]"
   (let [schema (strict (fnk-schema body-params))]
     (-> acc
         (update-in [:parameters :parameters] conj {:type :body :model schema})
         (update-in [:letks] into [body-params (src-coerce! schema :body-params :json)]))))
 
+; restructures form-params with plumbing letk notation. Example:
+; :form-params [id :- Long name :- String]
 (defmethod restructure-param :form-params [_ form-params acc]
-  "restructures form-params with plumbing letk notation. Example:
-   :form-params [id :- Long name :- String]"
   (let [schema (strict (fnk-schema form-params))]
     (-> acc
         (update-in [:parameters :parameters] conj {:type :form :model schema})
         (update-in [:letks] into [form-params (src-coerce! schema :form-params :query)]))))
 
+; restructures header-params with plumbing letk notation. Example:
+; :header-params [id :- Long name :- String]
 (defmethod restructure-param :header-params [_ header-params acc]
-  "restructures query-params with plumbing letk notation. Example:
-   :header-params [id :- Long name :- String]"
   (let [schema (fnk-schema header-params)]
     (-> acc
         (update-in [:parameters :parameters] conj {:type :header :model schema})
         (update-in [:letks] into [header-params (src-coerce! schema :headers :query)]))))
 
+; restructures query-params with plumbing letk notation. Example:
+; :query-params [id :- Long name :- String]
 (defmethod restructure-param :query-params [_ query-params acc]
-  "restructures query-params with plumbing letk notation. Example:
-   :query-params [id :- Long name :- String]"
   (let [schema (fnk-schema query-params)]
     (-> acc
         (update-in [:parameters :parameters] conj {:type :query :model schema})
         (update-in [:letks] into [query-params (src-coerce! schema :query-params :query)]))))
 
+; restructures path-params by plumbing letk notation. Example:
+; :path-params [id :- Long name :- String]
 (defmethod restructure-param :path-params [_ path-params acc]
-  "restructures path-params by plumbing letk notation. Example:
-   :path-params [id :- Long name :- String]"
   (let [schema (fnk-schema path-params)]
     (-> acc
         (update-in [:parameters :parameters] conj {:type :path :model schema})
         (update-in [:letks] into [path-params (src-coerce! schema :route-params :query)]))))
 
+; Applies the given vector of middlewares for the route from left to right
 (defmethod restructure-param :middlewares [_ middlewares acc]
-  "Applies the given vector of middlewares for the route from left to right"
   (assert (and (vector? middlewares) (every? (comp ifn? eval) middlewares)))
   (update-in acc [:middlewares] into (reverse middlewares)))
 
@@ -179,22 +177,36 @@
     `(-> (routes ~@body)
          ~@middlewares)))
 
-(defn- destructure-compojure-api-request [lets arg]
+(defmacro route-middlewares
+  "Wraps route body in mock-handler and middlewares."
+  [middlewares body arg]
+  (let [middlewares (reverse middlewares)]
+    `((-> (fn [~arg] ~body) ~@middlewares) ~arg)))
+
+(defn- destructure-compojure-api-request
+  "Returns a vector of three elements:
+   - new lets list
+   - bindings form for compojure route
+   - symbol to which request will be bound"
+  [lets arg]
   (cond
-    (vector? arg) [lets (into arg [:as +compojure-api-request+])]
+    ;; GET "/route" []
+    (vector? arg) [lets (into arg [:as +compojure-api-request+]) +compojure-api-request+]
+    ;; GET "/route" {:as req}
     (map? arg) (if-let [as (:as arg)]
-                 [(conj lets +compojure-api-request+ as) arg]
-                 [lets (merge arg [:as +compojure-api-request+])])
-    (symbol? arg) [(conj lets +compojure-api-request+ arg) arg]
+                 [(conj lets +compojure-api-request+ as) arg as]
+                 [lets (merge arg [:as +compojure-api-request+]) +compojure-api-request+])
+    ;; GET "/route" req
+    (symbol? arg) [(conj lets +compojure-api-request+ arg) arg arg]
     :else (throw
             (RuntimeException.
-              (str "unknown compojure destruction synxax: " arg)))))
+              (str "unknown compojure destruction syntax: " arg)))))
 
 (defn restructure [method [path arg & args]]
   (let [method-symbol (symbol (str (-> method meta :ns) "/" (-> method meta :name)))
         [parameters body] (extract-parameters args)
         [lets letks responses middlewares] [[] [] nil nil]
-        [lets arg-with-request] (destructure-compojure-api-request lets arg)
+        [lets arg-with-request arg] (destructure-compojure-api-request lets arg)
 
         {:keys [lets
                 letks
@@ -211,8 +223,8 @@
         body `(do ~@body)
         body (if (seq letks) `(letk ~letks ~body) body)
         body (if (seq lets) `(let ~lets ~body) body)
+        body (if (seq middlewares) `(route-middlewares ~middlewares ~body ~arg) body)
         body (if (seq parameters) `(meta-container ~parameters ~body) body)
         body `(~method-symbol ~path ~arg-with-request ~body)
-        body (if (seq middlewares) `(middlewares [~@middlewares] ~body) body)
         body (if responses `(body-coercer-middleware ~body  ~responses) body)]
     body))