+++ src/compojure/api/swagger.clj
@@ -1,21 +1,18 @@
 (ns compojure.api.swagger
-  (:require [clojure.string :as st]
-            [schema.core :as s]
+  (:require [clojure.set :refer [union]]
+            [clojure.string :as st]
+            [clojure.walk :as walk]
+            [compojure.api.common :refer :all]
+            [compojure.api.routes :as routes]
+            [compojure.core :refer :all]
             [plumbing.core :refer [fn->]]
-            [clojure.walk16 :as walk]
-            [clojure.set :refer [union]]
             [potemkin :refer [import-vars]]
-            [ring.util.response :refer :all]
+            [ring.swagger.common :refer :all]
             [ring.swagger.core :as swagger]
             [ring.swagger.impl :as swagger-impl]
-            [ring.swagger.common :refer :all]
             [ring.swagger.schema :as schema]
             ring.swagger.ui
-            [compojure.api.common :refer :all]
-            [compojure.api.core :as core]
-            [compojure.api.routes :as routes]
-            [compojure.route :as route]
-            [compojure.core :refer :all]))
+            [schema.core :as s]))
 
 ;;
 ;; Schema helpers