+++ test/compojure/api/sweet_test.clj
@@ -1,12 +1,12 @@
 (ns compojure.api.sweet-test
-  (:require [midje.sweet :refer :all]
-            [schema.core :as s]
-            [compojure.api.routes :as routes]
+  (:require [compojure.api.routes :as routes]
+            [compojure.api.sweet :refer :all]
+            [compojure.api.test-utils :refer :all]
+            [compojure.core :as compojure]
+            [midje.sweet :refer :all]
             [ring.mock.request :refer :all]
             [ring.swagger.schema :refer [describe]]
-            [cheshire.core :as cheshire]
-            [compojure.core :as compojure]
-            [compojure.api.sweet :refer :all]))
+            [schema.core :as s]))
 
 (s/defschema Band {:id s/Int
                    :name s/Str
@@ -105,7 +105,7 @@
 
   (fact "api-listing works"
     (let [{:keys [body status]} (api (request :get "/api/api-docs"))
-          body (cheshire/parse-string body true)]
+          body (parse-body body)]
       status => 200
       body => {:apiVersion "0.0.1"
                :apis [{:description "sample api"
@@ -115,6 +115,6 @@
 
   (fact "api-details works"
     (let [{:keys [body status]} (api (request :get (str "/api/api-docs/" app-name)))
-          body (cheshire/parse-string body true)]
+          body (parse-body body)]
       status => 200
       body => truthy)))